bot_service/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
aiogram==3.10.0
celery==5.4.0
redis==5.0.0
httpx==0.27.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiohttp==3.9.0
